+++ packages/router/src/utils.ts
@@ -11,6 +11,20 @@ export type AsyncCallback<T> = (done?: (t: T) => void) => Promise<T> | void
 export type SyncCallback<T> = () => T
 export type Callback<T> = AsyncCallback<T> | SyncCallback<T>
 export type MaybeArray<T> = T | T[]
+export type MaybePromise<T> = T | Promise<T>
+
+export function flatMap<T, R>(collection: T[], fn: (t: T) => MaybeArray<R>): R[] {
+  const flattened = []
+  for (const i of collection) {
+    const ret = fn(i)
+    if (Array.isArray(ret)) {
+      flattened.push(...ret)
+    } else {
+      flattened.push(ret)
+    }
+  }
+  return flattened
+}
 
 export async function sequence(callbacks: Callback<boolean | void>[], ...args: any[]): Promise<{
   count: number,
@@ -125,5 +139,9 @@ export const log = {
   error(...messages: string[]) {
     // tslint:disable-next-line no-console
     console.error('[@profiscience/knockout-contrib-router]', ...messages)
+  },
+  warn(...messages: string[]) {
+    // tslint:disable-next-line no-console
+    console.warn('[@profiscience/knockout-contrib-router]', ...messages)
   }
 }